# manager_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloakBrowser Manager API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/diagnostics": {
      "get": {
        "summary": "Diagnostics",
        "operationId": "diagnostics_api_v1_diagnostics_get",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "profile",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 36
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "direct_google_control",
                    "pixelscan",
                    "iphey",
                    "cloudflare",
                    "google_search"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "passed",
                    "warning",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticPage"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics/direct-google-control": {
      "post": {
        "summary": "Create Direct Google Control",
        "operationId": "create_direct_google_control_api_v1_diagnostics_direct_google_control_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/diagnostics/pixelscan": {
      "post": {
        "summary": "Create Pixelscan",
        "operationId": "create_pixelscan_api_v1_diagnostics_pixelscan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileDiagnosticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/diagnostics/iphey": {
      "post": {
        "summary": "Create Iphey",
        "operationId": "create_iphey_api_v1_diagnostics_iphey_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileDiagnosticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/diagnostics/cloudflare": {
      "post": {
        "summary": "Create Cloudflare",
        "operationId": "create_cloudflare_api_v1_diagnostics_cloudflare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileDiagnosticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/diagnostics/google-search": {
      "post": {
        "summary": "Create Google Search",
        "operationId": "create_google_search_api_v1_diagnostics_google_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileDiagnosticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/diagnostics/{diagnostic_id}": {
      "get": {
        "summary": "Diagnostic",
        "operationId": "diagnostic_api_v1_diagnostics__diagnostic_id__get",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnostic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics/{diagnostic_id}/artifacts/report": {
      "get": {
        "summary": "Report Artifact",
        "operationId": "report_artifact_api_v1_diagnostics__diagnostic_id__artifacts_report_get",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnostic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics/{diagnostic_id}/artifacts/screenshot": {
      "get": {
        "summary": "Screenshot Artifact",
        "operationId": "screenshot_artifact_api_v1_diagnostics__diagnostic_id__artifacts_screenshot_get",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnostic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics/{diagnostic_id}/cancel": {
      "post": {
        "summary": "Cancel Diagnostic",
        "operationId": "cancel_diagnostic_api_v1_diagnostics__diagnostic_id__cancel_post",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnostic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticRead"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Origin or CSRF rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosticPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "DiagnosticPage"
      },
      "DiagnosticRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "direct_google_control",
              "pixelscan",
              "iphey",
              "cloudflare",
              "google_search"
            ],
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "passed",
              "warning",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "target_url": {
            "type": "string",
            "enum": [
              "https://www.google.com/search?q=CloakBrowser+diagnostic",
              "https://pixelscan.net/",
              "https://iphey.com/",
              "https://challenge.cloudflare.com/turnstile/v0/generic/",
              "https://www.google.com/search?q=CloakBrowser+browser+diagnostic"
            ],
            "title": "Target Url"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Diagnostic completed.",
                  "Diagnostic completed with warnings.",
                  "Diagnostic failed.",
                  "Diagnostic cancelled."
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "findings": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "string",
                  "enum": [
                    "passed",
                    "warning",
                    "failed",
                    "unknown",
                    "aligned",
                    "mismatch",
                    "detected",
                    "not_detected",
                    "loaded",
                    "not_loaded",
                    "visible",
                    "not_visible",
                    "required",
                    "not_required"
                  ]
                }
              ]
            },
            "type": "object",
            "title": "Findings"
          },
          "screenshot_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screenshot Url"
          },
          "report_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Url"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "diagnostic_failed",
                  "manager_restarted",
                  "scheduler_unavailable",
                  "browser_crashed",
                  "cleanup_failed",
                  "profile_not_stopped",
                  "proxy_preflight_failed",
                  "network_error",
                  "timeout",
                  "target_layout_changed",
                  "captcha_user_action_required"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "The diagnostic could not be completed.",
                  "The manager restarted before the diagnostic completed.",
                  "The diagnostic could not be scheduled.",
                  "The browser closed before the diagnostic completed.",
                  "The diagnostic could not safely release its browser resources.",
                  "Stop the profile before running this diagnostic.",
                  "The assigned proxy is unavailable.",
                  "The diagnostic target could not be reached.",
                  "The diagnostic did not complete before its time limit.",
                  "The diagnostic target could not be read reliably.",
                  "The target requires user interaction."
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "profile_id",
          "kind",
          "status",
          "target_url",
          "requested_at",
          "started_at",
          "completed_at",
          "progress",
          "summary",
          "findings",
          "screenshot_url",
          "report_url",
          "error_code",
          "error_message"
        ],
        "title": "DiagnosticRead"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "field_errors": {
            "additionalProperties": true,
            "type": "object",
            "title": "Field Errors"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "ProfileDiagnosticRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Profile Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "profile_id"
        ],
        "title": "ProfileDiagnosticRequest"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cloak_session"
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    }
  }
}